Foundations for the cli rewrite
The idea is to include a very basic cli implementation within the library as binary.
For the moment, this is being built as an example, though in future xz support can be added as a feature and an explicit binary built using flags to enable support.
Still, for now, an example is sufficient.

The additional code should be fairly minimal, and should be useful as a reference.
At the moment luby_u is parsed and cores can be written.

With the FRAT updates, crossbeam is no longer needed, and so has been removed as a dependency.