Initial documentation for those things missing documentation
With the help of: warn(missing_docs)

There's not too much of interest, other than the documentation.
And, what is added could perhaps make do with a second pass, though for the most part documentation was overlooked at the items seem easy to understand.

A few redundant items have been removed.

A few functions have been renamed for clarity.