Minor watch list revisions
Binary and long watches have been split into distinct structs.

This simplifies a significant amount of pattern matching , though at the cost of some lost checks.

Still, with this methods to get and set watches have been split.
This allows the binary watch get to be const, rather than mut, which is an improvement.
It seems a pointer is still required, though, given the valuation may be mutated.
Still, intent is clearer.