Minor BCP improvements
The default implementation of literals has changed since the implementation of BCP.

Before, the boolean representation split an atom and the polarity of the atom.
Currently, the integer representation bundles both.

A handful of BCP methods took distinct atom and value parameters, derived from some literal bundle.
This revision keeps things bundled until a split is required.

My guess was that this wasn't an issue when working with the previous representation.
Though on reflection that's not obvious.
As, it's not obvious that fields of a struct are mapped to parameters, when possible.

In any case, this has a small but measurable benefit.