Minor update watch improvements
Incrementing the watch pointer is moved to after the first check.
In practice this means a witnessing watch will not be overlooked (intially).

Further, an assert has been added to ensure the procedure is safe, with an explanation.

And, the inline directive has been removed, which seems to help performance.

Elsewhere, queuing a literal has been given a functioning example.