Minor refinements to the resolution buffer
Set up for obtaining an asserting clause now uses the clause length to anticipate the appropriate capacity for the vec.

The 'asserts' field has been removed.
As it was, broadly speaking, not so useful.
The problem being, assertions were determined when a clause was merged with the buffer.
And, simply by checking whether there was no value for a literal.
And, as the buffer clears the current decision level, there may be multiple literals without a value, with all but one to be removed from the clause through resolution.

When obtaining an asserting clause, the asserted literal is in position zero.

The last is inspired by MiniSAT.
And, after looking back on this, I think the implementation in MiniSAT might be an improvement.

With the buffer, two full passes are made over the valuation.
The first to step back a decision, and the second to obtain the asserting clause.
This is significantly more than no passes in MiniSAT.

Though, on the other hand, the buffer does not need to keep track of which literals have been seen.
This, I think I recall, being the motivation for using a buffer.
As those checks vary.

Hm…