Slight tidy to BCP and related things
BCP now swaps in-loop and splits out-loop
This is slightly more efficient, though neglibigble.
The idea is due to MiniSAT implementing a similar operation.

On this, the implementation of BCP/updating watches in MiniSAT is quite similar to what is here.
The key differences are the absence of blocks here (as these were a clear hinderance in the tests ran) and use of a circular walk to update watches.
Though, what remains is mostly just filtering the watch list while walking through it.
On this point, though, the idea of maintaining order in the list returned, as MiniSAT does do this.
Equally, though, it may be more interesting to implement a heap.
Hm.

Elsewhere, minor changes.

More detailed notes on safety have been added to BCP.
Recording consequences has been touched up, as the result was mostly useless.
In place of an error on a missing clause, incrementing the proof count is skipped.

Valuations are no longer passed around when storing clauses, as this was unused and I doubt will have any use.

And, elsewhere slight formatting changes.