Simplification of the level stack
Consequences have been expanded to assignments, and now include assumptions and decisions.
And, with this, levels are now vectors of assignments.

The split between the decision and the consequences of that decision were not used.
So, little follows from this change.

- The separate methods to access decisions/consequences have been removed and replaced with a method returning all the assignments at a level.
- Places where the methods were called now call the unified method.

If required, the split is easy to infer, as decisions/assumptions are always prior to consequences.
Though, this is complicated a little by the possibility of multiple assumptions, and so taking on the first assignemtn may not be sufficient.
Still, this is for whenever the split may be of interest…