A revised level stack
Inspired by MiniSAT the level stack now uses marks on a single vec of assignments, rather than storing a vec of levels (which were vecs of assignments).
The code remains simple, and the perforamnce increase is clear.

There is little change to implementation, other than how levels are identified interally in the literal database.
Significant, only, is the removal of a method to gain mutable access to the top decision level, though this only had internal use (and it's now easier to simply push a new assignment to the list).