Intermediate commit
Attempting to resolve issues with assumptions.
And, things are messier than expected.

At present core tests pass ok, though IPASIR tests do not.
Further, while tests pass there is significant regression to performance.

At a high level the changes are:

- Solve now takes an optional collection of assumptions.
  + These are then used as before, but there's no longer the need to store anything.

- Backjump to a given level directly, through use of markers.
  + Most changes here are within the literal db.

- Revision of resolution with:
  + An updated track on how many literals are unassigned at the current level.
  + Laxer requirements on asserting clauses, specially to allow assertions of literals which have been valued.

- Given the revisions to resolution, solve has been updated to check the valuation of assertions, as the assertion itself may be sufficient for unsatisfaibility, and bcp does not begin until after the valuation has been made.

- A few additional logs.

The regression might be due to clearing too much from the trail.
Though, some efforts have been made to avoid this.

Roughly, the trail may have an initial segement which does not belong to a level.
This is fine, as whenever a level is created are mark is pushed and the level begins at the mark.
And, as levels are mapped to marks, level 0 is from the first mark…
Though, in practice this is unlikely to be relevant at present, as BCP without a decision or assumption transfers to the clause database.

In any case…

Issues with assumptions are almost certainly due to an incorrect initial_decision_level.
At the moment this points to the current level, which does not work in general.