At least one fix
Asserting flat assumptions now creates a level and then pushes all assumptions in the same loop.
For some reason this is significant.
All IPASIR tests of reasonable cost now pass.

Though, this is puzzling as I don't see a clear semantic difference in the code.

Unless, for some reason the tests were running on an outdated lib…