Returns for valuing/queuing simplified
Previously, the result from queuing collpased the result from valuing from three to two cases.
NotSet/Same were collapsed to an ok, and conflict to an error.
This added a lot of unnecesssary indirection, as did splitting the valuation result into ok/err.

Conceptually, too.
As, the same value is no issue in BCP, but arguably an issue when making an assumption.
So, the valuation prior to the attempted queue is returned for the caller of the queue operation, and the intermediate result has been removed.