Clause get_unchecked(_mut) simplified a little
The get_unchecked methods are designed to be used only when a clause is known to exist in the clause_db.
So, the previous half-checked with result has been revised closer to general unchecked.

In line with this, the use of get_unchecked_mut when removing clauses has been corrected to get.