Initial BCP, prior to asserting assumptions
In principle, this should help with iterative solves, as results which can be derived without an assumption will no longer be cleared with assumptions.

Otherwise, this does not have any significant impact.
Other than… highlighting the need for a refactor some time, as there are multiple exits to unsatisfiability, and all these require a callback, etc.
Still, some other time.