Improved failed literal accuracy
The last commit was almost correct…
Except for the identification of failed literals in basic tests.

To handle the case where an assumption directly conflicts with the valuation an additional case is required.
This amounts to looking up the source of the conflicting literal, whether original or addition.
For now, something of a hack, but will maybe be kept.

Really, though, assumptions need an ith pass…