Failed assumptions
The last commit introduced some functionality for handling conflicting assumptions which did not rely on BCP.

There I claimed all standard tests pass.
Which, was not quite right.
The additional code to identify the conflict failed some IPASIR tests, as I assumed tests would only fail due to some original or addition unit and did not retest.

Hek.

Well, now there's no panic and a conflict with top is returned on a failed assumption without an identified original/additon unit.

And, generally speaking, it's probably not a good idea to panic on an awaited design decision for something which I expect would only occur due to misconfiguration --- though perhaps this hints that some issue remains… as conflict without a unit… in any case, this commit amends the previous and tests do now pass.