Small refinements, relocated tests
Added:

- A method to obtain a vector of fresh/max literals.
- A method to propagate the queue, or return an error on unsat.
  In other words, the core of solve without decisions or backtracking, but making those parts of solve conditional would be messier than maintaining both methods.
- A method to unify noting a conflicting, making the callback, etc.

Tests which do not rely on external CNF files have been moved from the tests crate to the main crate and have been simplified a little.
And, some basic tests for assumptions have been added.

Some small refinements to the initial part of solve have been made.
In particular, the initial BCP instance now handles verifying the CNF is satisfiable before the core loop, and so parts of the assumption code are less convoluted.