Quick replacement of implicit res. graph with map
Previously the implication graph used to identify an unsat core was kept as part of the clause database.

Specifically, each use of a clause in resolution added to the inference count of a clause, and clauses were only removed if the count was zero.

Otherwise expressed, reference counting was applied to clauses with respect to their appearance in the resolution graph, and only clauses with a count of zero were allowed to be removed.

This was intended as an temporary experimental approach to deriving unsatisfiable cores.

Still, there's not really much to experiment with.
There's no need to store anything other than the graph as given through keys.
For, the only clauses of interest are those which appear in the original formula.
And, both intuition and experimentation suggest that the rigidity imposed by removing only leaves from the graph when removing clause from the database is a burden.
In particular, on the standard tests there's a signficant improvement when the rigidity is relaxed.

This commit (simply) removes the reference counting and replaces the implicit graph with a hash map from keys to sets of keys, exactly those keys stored in clauses prior.

Of note is that updating the map can be centrailised to the addition of a cluase, which has been done.

Remaining is to tidy up and verify cores are correctly identified.