Initial merge of atom and literal dbs
Initially, the literal db has been merged into the atom db.
Though, the goal is to recast the atom db as the 'assignment db', e.g.

The practical motivation, in short, is that this involves less passing around of databases as parameters.
Conceptually it makes sense, too, as there's a clean distinction between the db of clauses used to make an assignment and the assignment itself, whether as a valuation, trail, etc. (though that's not to say further changes to the db won't follow).

With respect to functionality there are two changes:

- Dropping assignments when backjumping has been merged into the clear assignment method.
  Primarily, required due to multiple mutable borrows.
- Subsumption has been temporarily disabled.
  Again, due to mutable borrows, and deferred as MiniSAT has shown I should revise the ideas regardless.