Clause addition revisions
Checked and unchecked clause addition methods have been merged, in favour of unchecked.

In the context of prior commits, the goal is to revise assignments to keep track of unit clauses, which in turn will ease finding cores.

Specific to this commit, two methods are difficult to maintain, and the role of checked addition can be simulated by first checking whether the clause would be unsatisfiable on the given valuation prior to (attempting) to add the clause.

This led to some issues with the IPASIR API, though primarily due to issues with conflicting assumptions, and so recording conflicts has been bodged, for the moment.
Similarly, a doc test which relied on checked addition has been amended (though may no longer make as much sense in the context of the docs).