Queing paired to assignment
The current goal is to refine assignments.
And in particular so that cores are easy to identify.

This commit is prep work for merging the consequence queue into the assignment trail.

Overall, the observation is that there's really no need for a distinct queue.
Instead, it'll be sufficient to keep a marker on which assignments have been propagated.

To help set this up, each use of value_and_queue has been paired to an instance of store_assignment (before, store_consequence).
In the furture, value_and_queue can then be replaced with a value check.

And, additional assignment sources for unit original/addition clauses have been added.