'Direct' valuation of atoms
Without the q, value_and_queue amounted to unnecessary indirection.
For the moment, this indirection has been removed.
The result is a slight increase to complexity when calling a valuation, though this is preferable for the moment.

In addition, the atom db docs have been updated slightly, with notes on the queue.
As such, the q file has been removed.