Revised core generation, and tests
First pass

Core identification is inspired by MiniSAT, specifically the code for failed assumptions.
Though, inspiration at this point means the idea has sat for a week or so, and does not build directly on the source.
So, a few more passes are required.

Simple tests for the cores have been added.
The simply verify the unsat core with a new context.
These may be built out further later.

For the moment assumptions are omitted from an unsat core.
This is something to be revisited…