Subsumption restored, minor tidying
Subsumption is enabled by providing methods to take/restore the trail.
This is safe to do in the context of resolution, where used, as the resolution cannot update the trail.

In addition, as subsumption no longer transfers clauses, the indirection of the clause_db subsumption method has been removed.

Elsewhere, documentation has been updated a little.
Though, parts remain quite out of date.