Refinements to resolution buffer cells
The current goal is to revise the resolution buffer to store information about BCP.
This, in particular, will allow for improvements to subsumption, following MiniSAT.

The BCP information will be used to identify redundant literals in a conflict clause, similar to MiniSAT.
And, an alterative would be to store these in an additional 'reason' vector.
Still, mutations to the resolution buffer happen only during resolution, together with mutations to the resolution buffer.
So, it seems reasonable expand the buffer.

This initial commit replaces stored literals with boolean values, ensuring the index to a cell corresponds to an atom.