Further resolution buffer refinements
A 'used_stack' has been introduced to track the literals used.
In other words, the literals which appeared in some merged clause.

The key upshot is avoiding traversal of the full buffer in order to identify clauses, etc.
In particular, updates to the buffer are now made together with assignments and parts of the buffer are cleared when backjumping.

Likewise for bumping used atoms.
Though, with the inclusion of this change performance is difficult to measure as there is likely an impact to heuristics.
(Specifically, as bumping happens in a different order, normalisation is impacted.)

Generally this seems ok.