Initial extension of cells to contain assignment information
A struct replaces the previous enum, with the relevant options of the enum reformed as a field.

It might be preferable to switch optional values so that cells are an extension of assignments, though there's some headache either way.
Hm, though on continued reflection, that would allow unification of assignment information…

Otherwise, no changes.