Transfer of assignments to the resolution buffer
The trail has been cut to only track literals.
Unused methods relating to assignments have been commented out, and uses of assignments now pass through the resolution buffer.

The previous treatment of asserting assumptions required revision, as it relied on identifying assumptions through assignments, and with the change no atom can be assigned multiple times (e.g. as an assumption and as a conflict).
Though, an easy fix is applied by identifying assumptions through their position in the trail.