Refinements to the atom db
At present, the broad goal is to refactor the context to support consolidating a few distinct structures.
The immediate goal is to --- roughly --- dissolve the atom db and raise things to the context.
The primarily benefit is to flatten the structure of a context to avoid the current difficulty with moving things around.

In this commit:

- Fields of the atom db related to the trail have been consolidated to a fresh 'trail' struct.
- Various methods which related to the trail have been associated with the struct.

- The fundamental way to obtain an atom has been raised to the context.
- With this, initialisation of 'the true' is no longer possible when creating the atom db, so an init method has been added to the generic context (perhaps temporarily).