Initial raise of watches to the context
Rough, for the moment.

Minor changes as required.
And, to keep things simple the watch dbs are wrapped in a Watches struct, with gets/sets of watches as methods to the struct.

Aside from different access routes, the only change is passing watches separately when needed through some methods.
As these cases already handled the atom db no refactoring was required.