Atom activity raised, and syntax
Resolution buffer renamed to atom cells…

well… imagination is difficult for the moment…

Otherwise, methods from atom db valuation moved to mod and the file removed.

A miskey and this commit also raises the index heap for atom activity to the context.
In order to handle borrows the methods for manipulating the heap have been recast as functions.
Ideally these can be generalised as indexheap methods, or at least to clauses.
Still, for now they allow the raise.

With this the atom config has also been removed.