Proven literal skipped when learning, core bugfix
Proven literals are safely skipped during resolution (and with benefits to clause length, etc.).

As whether a literal has been proven is given when the assignment is made and the resolution status is cached, ResolutionStatus has been expanded to with a case for proven literals.
The case is set when making an assignment, and remains set.

This information is available by checking the decision level of a literal, which is also present in a cell.
Still, the enum case allows for some nice matching.

Elsewhere, minor syntactic changes.