Subsumption, following MiniSAT (initial pass)
For the moment, the use of subsumption is a hardcoded boolean and ideally some revisions are possible.

Still, an initial pass with functional subsumption makes sense.

In general, the way subsumption is handled is the same as in MiniSAT.
That is, iterative DFS through BCP history, with tags set to cache work done.

A significant departure, and issue, is that binary clauses do not maintain the --- asserted literal in position 0 --- invariant.
This invariant has extensive implicit use in the MiniSAT approach, and so to adapt the approach JIT maintenance of the invariant is implemented.
Attempts were made to avoid this, but for an initial pass things became much too complex.

A few supporting methods have been added to clauses to help simplify subsumption a little.