Config for strengthening
For the moment, only to enable recursive strengthening.

It would be nice to have a toggle for proven literals.
However, the implementation rests on cell status, and I don't think this should be bound to strengthening (even if it has no other role at the moment).
So, addition would require further cell discriminants, and this doesn't seem sufficiently interesting for now.