Refined recursive strengthening
The previous appraoch to recusrive strengthening performed JIT maintenance of the zero-index-is-asserted invariant for binary clauses.
This isn't required.
For, in the case of binary clauses, only one literal requires inspection, and so instead of maintaining the invariant it's possible to introduce specialised updates for the index.
That is, rather than always initialising at 1 and incrementing, intiailise to either 0 or 1 and on what would be an increment jump to 2.
As, then, the correct literal is inspected and the clause is correctly noted as exhausted when backtracking in the DFS, which is all that's required.

The above has been implemented, and overall the strengthening method has been refined for clarity.