Minimization for proven only, misc updates
A new minimization option to omit proven literals only is implemented.
This works by keeping the proven status/flag as before, but only setting the flag when proven literals should be omitted.

Removing the proven status/flag in favour of checks on the level of an atom was explored, but all the additional checks added significant overhead.

For clarity, CellStatus has been resyntaxed to ResolutionFlag.
As, following the notes in a previous commit, the field should clearly be for resolution, or independent.
And, in contrast to the previous commit, I don't see much interest if independent from resolution.

Elsewhere, the derivable check has been refined a little more for clarity.
Of note, the stack of tuples now uses a custom struct.
And, some duplicate code has been merged/replaced by a function call.