Removed atom module from db module
The module was no longer relevant, with the contained items split as follows:

activity_bump -> Merged with the method for bumping clause activity in an 'activity.rs' file in the db module.
ValuationStatus enum -> Moved to the valuation module in structures

Aside:

On reflection it doesn't make too much sense to merge acitivty methods.
As, there's some significant divergence.
Of note, a top atom always exists, but a top clause need not…