Additional IPASIR fix
At some point handling of an assumption with an existing value was revised.
In the revision, any existing value meant the assumption failed to be set.

This makes sense to me, and no doubt is why it was set this way.

Still, the IPASIR API handles things differently.
An attempt to assume a literal which has already be established is handled as the assumption holding.

I can also see why this is nice.

With this revision reverted, all non-expensive IPASIR tests pass.